requirements/requirements.txt
Scrapy==2.6.2
pyOpenSSL==22.0.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.5 # psycopg2==2.9.3
pika==1.3.0
sqlalchemy-migrate==0.13.0 